+++ lib/screens/hunting_field_screen/zones/enemy_page/enemy_page.dart
@@ -46,9 +46,10 @@ class EnemyPage extends StatefulWidget {
   State<EnemyPage> createState() => _EnemyPageState();
 }
 
-class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMixin {
+class _EnemyPageState extends State<EnemyPage>
+    with SingleTickerProviderStateMixin {
   bool _showDropList = false;
-  late int _enemyCurrentHP;
+  late double _enemyCurrentHP;
   final GlobalKey<AnimatedListState> _listKey = GlobalKey<AnimatedListState>();
   final List<Item> _dropHistory = [];
 
@@ -75,7 +76,7 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
   @override
   void initState() {
     super.initState();
-    _enemyCurrentHP = widget.enemy.hp;
+    _enemyCurrentHP = widget.enemy.hp.toDouble();
 
     _pulseController = AnimationController(
       vsync: this,
@@ -119,7 +120,8 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
     if (_isCombatActive) return;
     _isCombatActive = true;
     final intervalMs = (widget.enemy.attackSpeed * 1000).toInt();
-    _enemyAttackTimer = Timer.periodic(Duration(milliseconds: intervalMs), (timer) {
+    _enemyAttackTimer =
+        Timer.periodic(Duration(milliseconds: intervalMs), (timer) {
       _enemyAttack();
     });
   }
@@ -139,14 +141,14 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
     final charState = characterBloc.state as CharacterLoaded;
     if (charState.character.currentHealth <= 0) return;
 
-    int maxDamageBlocked = (widget.enemy.attackDamage * 0.8).toInt();
-    int actualDamageBlocked = charState.character.defense;
+    double maxDamageBlocked = widget.enemy.attackDamage * 0.8;
+    double actualDamageBlocked = charState.character.defense.toDouble();
 
     if (actualDamageBlocked > maxDamageBlocked) {
       actualDamageBlocked = maxDamageBlocked;
     }
 
-    int damageTaken = widget.enemy.attackDamage - actualDamageBlocked;
+    double damageTaken = widget.enemy.attackDamage - actualDamageBlocked;
     if (damageTaken < 0) damageTaken = 0;
 
     characterBloc.add(CharacterTakeDamage(widget.enemy.attackDamage));
@@ -165,7 +167,8 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
   void _startPlayerAttack() {
     _isWeaponOnEnemy = true;
     _tryPlayerAttack();
-    _playerAttackTimer = Timer.periodic(const Duration(milliseconds: 16), (timer) {
+    _playerAttackTimer =
+        Timer.periodic(const Duration(milliseconds: 16), (timer) {
       if (!_isWeaponOnEnemy) {
         timer.cancel();
         return;
@@ -222,10 +225,12 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
     });
 
     if (_enemyCurrentHP <= 0) {
-      _enemyCurrentHP = widget.enemy.hp;
+      _enemyCurrentHP = widget.enemy.hp.toDouble();
 
       if (widget.enemy.expReward > 0 || widget.enemy.spReward > 0) {
-        context.read<CharacterBloc>().add(CharacterAddExp(widget.enemy.expReward, spAmount: widget.enemy.spReward));
+        context.read<CharacterBloc>().add(CharacterAddExp(
+            widget.enemy.expReward,
+            spAmount: widget.enemy.spReward));
       }
 
       final loot = LootService.generateLoot(widget.enemy);
@@ -238,7 +243,8 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
           inventoryBloc.add(InventoryEvent$AddItem(item: item));
         }
         _dropHistory.insert(0, item);
-        _listKey.currentState?.insertItem(0, duration: const Duration(milliseconds: 300));
+        _listKey.currentState
+            ?.insertItem(0, duration: const Duration(milliseconds: 300));
       }
     }
   }
@@ -246,10 +252,14 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
   String _getItemName(Item item) {
     if (item is GoldItem) return '${item.amount} Gold';
     if (item is Weapon) {
-      return item.enchantLevel > 0 ? "${item.name} +${item.enchantLevel}" : item.name;
+      return item.enchantLevel > 0
+          ? "${item.name} +${item.enchantLevel}"
+          : item.name;
     }
     if (item is Armor) {
-      return item.enchantLevel > 0 ? "${item.name} +${item.enchantLevel}" : item.name;
+      return item.enchantLevel > 0
+          ? "${item.name} +${item.enchantLevel}"
+          : item.name;
     }
     if (item is Scroll) return item.name;
     return 'Unknown Item';
@@ -269,7 +279,8 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
   }
 
   void _onWeaponSlotCollapsed() {
-    final escapeTime = context.read<CharacterBloc>().state.character.escapeCooldownEndTime;
+    final escapeTime =
+        context.read<CharacterBloc>().state.character.escapeCooldownEndTime;
     if (escapeTime == null || DateTime.now().isAfter(escapeTime)) {
       setState(() {
         _isWeaponSlotExpanded = false;
@@ -321,20 +332,24 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
     return BlocListener<CharacterBloc, CharacterState>(
       listenWhen: (previous, current) {
         if (previous is CharacterLoaded && current is CharacterLoaded) {
-          return current.character.currentHealth != previous.character.currentHealth;
+          return current.character.currentHealth !=
+              previous.character.currentHealth;
         }
         return false;
       },
       listener: (context, state) {
         if (state is CharacterLoaded) {
-          final healAmount = state.character.currentHealth - _previousPlayerHealth;
+          final healAmount =
+              state.character.currentHealth - _previousPlayerHealth;
           if (healAmount > 0 && _previousPlayerHealth != -1) {
             final id = UniqueKey().toString();
             setState(() {
               _playerDamageTexts.add(DamageTextData(
                 id: id,
-                damage: healAmount,
-                randomX: _isWeaponSlotExpanded ? (_random.nextDouble() * 40 - 20) : 0,
+                damage: healAmount.toDouble(),
+                randomX: _isWeaponSlotExpanded
+                    ? (_random.nextDouble() * 40 - 20)
+                    : 0,
                 randomY: _random.nextDouble() * 40 - 20,
                 isHeal: true,
               ));
@@ -355,7 +370,10 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
           return Container(
             decoration: const BoxDecoration(
               color: AppColors.panelBackground,
-              image: DecorationImage(image: AssetImage('assets/background/forest_enemy_background.png'), fit: BoxFit.cover),
+              image: DecorationImage(
+                  image: AssetImage(
+                      'assets/background/forest_enemy_background.png'),
+                  fit: BoxFit.cover),
             ),
             width: widget.width,
             child: LayoutBuilder(
@@ -402,7 +420,8 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
                                       damage: dt.damage,
                                       flyUp: false,
                                       isHeal: dt.isHeal,
-                                      onComplete: () => _onPlayerDamageTextComplete(dt.id),
+                                      onComplete: () =>
+                                          _onPlayerDamageTextComplete(dt.id),
                                     ),
                                   ),
                                 );
@@ -423,7 +442,8 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
                           child: EnemyHpBar(
                             width: widget.width,
                             enemy: widget.enemy,
-                            widthOfOneHP: (widget.width - 200) / widget.enemy.hp,
+                            widthOfOneHP:
+                                (widget.width - 200) / widget.enemy.hp,
                             currentHP: _enemyCurrentHP,
                             heightFactor: enemyHpHeight,
                           ),
@@ -458,7 +478,8 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
                                         child: DamageTextWidget(
                                           damage: dt.damage,
                                           flyUp: true,
-                                          onComplete: () => _onEnemyDamageTextComplete(dt.id),
+                                          onComplete: () =>
+                                              _onEnemyDamageTextComplete(dt.id),
                                         ),
                                       ),
                                     );
@@ -509,17 +530,27 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
                             child: Column(
                               mainAxisSize: MainAxisSize.min,
                               children: widget.enemy.dropList.map((drop) {
-                                final item =
-                                    ItemRegistry.createItem(drop.itemType, weaponType: drop.weaponType, armorType: drop.armorType, scrollType: drop.scrollType);
+                                final item = ItemRegistry.createItem(
+                                    drop.itemType,
+                                    weaponType: drop.weaponType,
+                                    armorType: drop.armorType,
+                                    scrollType: drop.scrollType);
                                 String titleText = _getItemName(item);
                                 if (item is GoldItem) {
-                                  titleText =
-                                      drop.minQuantity == drop.maxQuantity ? '${drop.minQuantity} Gold' : '${drop.minQuantity}-${drop.maxQuantity} Gold';
+                                  titleText = drop.minQuantity ==
+                                          drop.maxQuantity
+                                      ? '${drop.minQuantity} Gold'
+                                      : '${drop.minQuantity}-${drop.maxQuantity} Gold';
                                 }
                                 return ListTile(
-                                  leading: Image.asset(item.image, width: 40, height: 40, gaplessPlayback: true),
-                                  title: Text(titleText, style: AppTypography.bodyMediumPrimary),
-                                  trailing: Text('${drop.chance}%', style: AppTypography.bodyLargeHighlight),
+                                  leading: Image.asset(item.image,
+                                      width: 40,
+                                      height: 40,
+                                      gaplessPlayback: true),
+                                  title: Text(titleText,
+                                      style: AppTypography.bodyMediumPrimary),
+                                  trailing: Text('${drop.chance}%',
+                                      style: AppTypography.bodyLargeHighlight),
                                 );
                               }).toList(),
                             ),
@@ -538,7 +569,8 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
                               children: [
                                 Text(
                                   'YOU ARE DEAD',
-                                  style: AppTypography.titleLargePrimary.copyWith(
+                                  style:
+                                      AppTypography.titleLargePrimary.copyWith(
                                     color: AppColors.error,
                                     fontSize: 48,
                                     fontWeight: FontWeight.bold,
@@ -555,7 +587,8 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
                                 Text(
                                   'You lost:\n-25% exp\n-50% gold\n-25 SP\n\nHunting is now unavailable for 15 seconds.',
                                   textAlign: TextAlign.center,
-                                  style: AppTypography.titleMediumPrimary.copyWith(
+                                  style:
+                                      AppTypography.titleMediumPrimary.copyWith(
                                     color: AppColors.enemyHp,
                                     fontWeight: FontWeight.bold,
                                   ),
@@ -565,15 +598,20 @@ class _EnemyPageState extends State<EnemyPage> with SingleTickerProviderStateMix
                                   onPressed: _onRebornPressed,
                                   style: ElevatedButton.styleFrom(
                                     backgroundColor: AppColors.success,
-                                    padding: const EdgeInsets.symmetric(horizontal: 40, vertical: 20),
+                                    padding: const EdgeInsets.symmetric(
+                                        horizontal: 40, vertical: 20),
                                     shape: RoundedRectangleBorder(
                                       borderRadius: BorderRadius.circular(12),
-                                      side: BorderSide(color: AppColors.primaryText.withValues(alpha: 0.24), width: 2),
+                                      side: BorderSide(
+                                          color: AppColors.primaryText
+                                              .withValues(alpha: 0.24),
+                                          width: 2),
                                     ),
                                   ),
                                   child: Text(
                                     'REBORN',
-                                    style: AppTypography.titleLargePrimary.copyWith(
+                                    style: AppTypography.titleLargePrimary
+                                        .copyWith(
                                       fontWeight: FontWeight.bold,
                                     ),
                                   ),
